- changed folder structure - Updated icon - Removed the export from __gen__
