+++ ui/codegen.ts
@@ -5,7 +5,7 @@ const pluginsDirectory = `${rootGQLDirectory}/plugins`;
 
 const config: CodegenConfig = {
   schema: "../api/gql/*.graphql",
-  documents: ["src/lib/gql/queries/**/*.ts"],
+  documents: ["src/lib/gql/**/*.ts"],
   ignoreNoDocuments: true, // for better experience with the watcher
   generates: {
     [rootGQLDirectory]: {
