+++ ui/src/lib/gql/queries/index.ts
@@ -1,2 +1,1 @@
 export * from "./user";
-export * from "./workspace";
