+++ ui/src/lib/gql/workspace/index.ts
@@ -1,0 +1,2 @@
+export * from "./hooks";
+export * from "./types";
