+++ ui/src/lib/gql/workspace/types.ts
@@ -1,0 +1,8 @@
+import type { Member, Project } from "@flow/types";
+
+import type { Workspace as WorkspaceGraphqlType } from "../__gen__/graphql";
+
+export type Workspace = Pick<WorkspaceGraphqlType, "id" | "name" | "personal"> & {
+  members?: Member[];
+  projects?: Project[];
+};
